# KeyAnalyzer tests and small tweaks.