Hoisted the calculation of the maximum collisions allow outside the
HasSufficientUniquenessFactor as the calculation only needs
to be done once (also makes it more testable).

Hoisted the calculation of the maximum offset outside of the offset
loops (as it's only dependent on the slice count and minimum length).

Made ContainsAnyLetters and SufficientUniquenessFactor
internal so they can be tested.

Added tests for ContainsAnyLetters and SufficientUniquenessFactor are working correctly.