@stephentoub 2023-08-10
> Benchmarks are identical

Can we please undo the hoisting-related changes then?  Let's keep this just tests.