@IDisposable 2023-10-19
Yep, that was from when it was still using my other `KeyAnalyzer` changes (that resulted in an empty set when `HasSufficientUniquenessFactor` returned `false`). 

Since we're not using that code right now, the test was wrong, sorry.

Thanks for fixing it.